+++ packages/interfaces-adt/src/adt/IAdtInclude.ts
@@ -30,6 +30,6 @@ export interface IIncludeConfig {
   packageName?: string; // Required for create operations, optional for others
   transportRequest?: string; // Only optional parameter
   description?: string; // Required for create/validate operations, optional for others
-  sourceCode?: string;
+  source?: string;
   onLock?: (lockHandle: string) => void;
 }