+++ packages/interfaces-adt/src/adt/IAdtInterface.ts
@@ -13,7 +13,5 @@ export type IInterfaceConfig = IAdtObjectConfig & {
   packageName?: string; // Required for create operations, optional for others
   transportRequest?: string; // Only optional parameter
   description?: string; // Required for create/validate operations, optional for others
-  sourceCode?: string;
-  sessionId?: string;
-  onLock?: (lockHandle: string) => void;
+  source?: string;
 };